+++ packages/DataElement/src/index.ts
@@ -1,0 +1,7 @@
+export {
+    customDataElements,
+    DataElement
+} from "./DataElement";
+export { 
+    StateChange
+} from "@domx/statechange";