+++ packages/Middleware/CHANGELOG.md
@@ -4,13 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.9] 2021-09-22
+### Updated
+- dependencies and audit
+
+## [0.2.8]  - 2021-09-10
+### Updated
+- Documentation 
+
+## [0.2.6] - 2021-08-05
+### Updated
+- Updates for package publishing
+
 
 ## [0.2.2]
 ### Added
 - A clean method
 
 ## [0.2.0]
-### TODO 
 ### Added
 - Middleware implementation
 - Middleware tests and documentation